scripts/archive/requirements.txt
supabase>=2.5.0
python-dotenv
requests
aiohttp
edge-tts
pillow
google-genai